+++ impl/src/as/mod.rs
@@ -239,7 +239,9 @@ impl<'a> ToTokens for Expansion<'a> {
                 ImplKind::Specialized
             };
 
-            let trait_ty = quote! { ::core::convert::#trait_ident <#return_ty> };
+            let trait_ty = quote! {
+                ::derive_more::#trait_ident <#return_ty>
+            };
 
             let generics = match &impl_kind {
                 ImplKind::Forwarded => {
@@ -251,7 +253,7 @@ impl<'a> ToTokens for Expansion<'a> {
                     if is_blanket {
                         generics
                             .params
-                            .push(parse_quote! { #return_ty: ?::core::marker::Sized });
+                            .push(parse_quote! { #return_ty: ?::derive_more::core::marker::Sized });
                     }
                     Cow::Owned(generics)
                 }
@@ -272,7 +274,7 @@ impl<'a> ToTokens for Expansion<'a> {
 
                     let conv =
                         <::derive_more::__private::Conv<& #mut_ #field_ty, #return_ty>
-                         as ::core::default::Default>::default();
+                         as ::derive_more::core::default::Default>::default();
                     (&&conv).__extract_ref(#field_ref)
                 }),
             };