+++ impl/src/not_like.rs
@@ -36,7 +36,7 @@ pub fn expand(input: &DeriveInput, trait_name: &str) -> TokenStream {
 
     quote! {
         #[automatically_derived]
-        impl #impl_generics ::core::ops::#trait_ident for #input_type #ty_generics #where_clause {
+        impl #impl_generics ::derive_more::#trait_ident for #input_type #ty_generics #where_clause {
             type Output = #output_type;
 
             #[inline]
@@ -108,7 +108,7 @@ fn enum_output_type_and_content(
                 let method_iter = method_iter.by_ref();
                 let mut body = quote! { #subtype(#(#vars.#method_iter()),*) };
                 if has_unit_type {
-                    body = quote! { ::core::result::Result::Ok(#body) }
+                    body = quote! { ::derive_more::core::result::Result::Ok(#body) }
                 }
                 let matcher = quote! {
                     #subtype(#(#vars),*) => {
@@ -135,7 +135,7 @@ fn enum_output_type_and_content(
                     #subtype{#(#field_names: #vars.#method_iter()),*}
                 };
                 if has_unit_type {
-                    body = quote! { ::core::result::Result::Ok(#body) }
+                    body = quote! { ::derive_more::core::result::Result::Ok(#body) }
                 }
                 let matcher = quote! {
                     #subtype{#(#field_names: #vars),*} => {
@@ -147,7 +147,7 @@ fn enum_output_type_and_content(
             Fields::Unit => {
                 let operation_name = method_ident.to_string();
                 matches.push(quote! {
-                    #subtype => ::core::result::Result::Err(
+                    #subtype => ::derive_more::core::result::Result::Err(
                         ::derive_more::UnitError::new(#operation_name)
                     )
                 });
@@ -162,7 +162,7 @@ fn enum_output_type_and_content(
     };
 
     let output_type = if has_unit_type {
-        quote! { ::core::result::Result<#input_type #ty_generics, ::derive_more::UnitError> }
+        quote! { ::derive_more::core::result::Result<#input_type #ty_generics, ::derive_more::UnitError> }
     } else {
         quote! { #input_type #ty_generics }
     };